Refactor game loop (#25)
* Refactor game loop
- Moved game loop inside `Game`
- Encapsulated fps logic inside `TickerTime` object
- Refactored the game loop logic, so it loops in a constant matter (even when update is faster)

* Change to float parameter of Update

* Refactor InputManager so it uses Points

* Refactor player movement with Point structure

* Let flocks also make use of the destination property

* Let the bullet use destination property
- also increase speed

* Reduce flocking size for stable enviroment
- pc can't handle 100 :(

* Add high res ticks and replace SDL_Delay
- Replace SDL_Delay with own "busy" loop (to force draw only at given FPS). As SDL_Delay is not relayable on different systems
- Use more accurate tick getter

* Cleanup

+++ Game/PlayingState.h
@@ -15,7 +15,7 @@ public:
 
     void HandleEvents(Game &game) override;
 
-    void Update(Game &game, int time) override;
+    void Update(Game& game, float time) override;
 
     void Draw(Game &game) override;
